tests are corrected
